Added output code
Operator will now export a formatted python file to a directory of your choosing. The file is formatted to run in Unreal, and recreate the material as it appears in Blender.
So far, it will create the material with the proper name, in the location provided, but it will not yet add nodes (working on it).