Bug Fixes
Fixed some small bugs in logic where i used NOT - OR, instead of NOT - AND NOT, fixed a few keys in the socket_translate dictionary (eg Separate XYZ has Vector Output, Separate RGB and HSV have Image Output).
Found a serious flaw in the code: attempting to port the Mapping node over (which contains lots of subtract, multiply, sine etc nodes) I started getting 0071, and 0090 as KeyErrors, when the code should only ever return a 0 or 1. Turns out, since all operations in Blender are handled by one node - Math - and all inputs have only one name - Value - that all the inputs are labelled sequentially throughout the material, rather than just within a single node.

For example, if we have one Math node, it will have input 0 and input 1, but, if we add another, it will be given input 2 and input 3, and so on, such that I can't seem to get the proper key to return the value I need from the socket_translate dictionary.

I put a little code in, in theory, I should be able to get the number I need by saying: socket = socket_number - 2 * node_ID_number, but it's still giving me key errors, so there's something up with my logic somewhere. bleh. As far as I've tested, this isn't impacting simple materials, only materials with multiple instances of the same Math operation.