Multiple Math Node Support Added
In my testing, I found that the actual magic equation was: socketnumber - linktonodeID * 10, and it gave me the results I was after.
I can now successfully reproduce PGMath's reproduction (say that three times fast) of Blender's Mapping node in Unreal Engine.
This and ColorRamp will make up the 0.4 release.
My intention is to include a Boolean Property (checkbox) above the 'Transfer Material' button, called 'First Time Setup Script' or something, that will include the Mapping Node as a Material Function, as well as the ColorRamp, which will be subdivided into several different Material Functions, to be called dependent on the number of colours used in the ColorRamp (eg mat_func_2_color_ramp, mat_func_3_color_ramp, mat_func_4_color_ramp etc). 
This should expand the functionality sufficiently for wider testing, and, after that, it'll be a matter of figuring out which nodes to prioritize.