Improved readability, added comments
Removed a few blocks of code left over from experiments, neatened up the texture names, they should be a little shorter, and fall in line with the node naming convention.
Added more comments throughout to explain what is happening at each stage. The idea is to make it full of long, descriptive variable names, and lots of comments that tell what is happening to those variables in such a way that someone with little knowledge of python could still pick it up relatively quickly.
All experiments are currently being done on the (complex) PGMath Mapping material, and it all seems to be holding up.