Fixed: Reroute KeyError
Caught this bug as I was testing the ColorRamp transfer. Unreal's Reroute has an input and output that are both nameless, I had assumed Blender's was the same, but it turns out they're called "Input" and "Output", who could have imagined?
In other news, the Mapping node has been recreated as a Material Function in Unreal, the Color Ramp is partway through that same process. When they're finished, and formatted into a startup script, that will be 0.4, or, possibly Mapping will be 0.4 and ColorRamp will be 0.5, depending on the thorns I encounter on the way.