ColorRamps now exposed to Material Library
Small change, the ColorRamp Material Functions now show up in the sidebar of the Material editor in Unreal, so you don't need to folder dive to find them, you can just drag and drop. I added them to a new category, "BLUI", as well as "Custom" and "Utility". They could go in more or less, probably just having them in there once is enough, I can adjust based on user feedback.
All other limitations with ColorRamps still apply. The "post load nodes" pass is looking more and more like the solution, and the only thing preventing it from happening is figuring out how to get the correct socket data for all the new nodes that the additional pass would create.
If this approach was taken, it could solve the "typing values into Principled BSDF not transferring" issue, to boot.