Post Load Nodes and better ColorRamp support
Implemented a second pass of node creation to aid in transferring ColorRamp settings. In Blender, you can manipulate all the settings directly in your material. In Unreal, you would either have to create the entire function in your material (messy, lots of nodes) or double click to edit the material function to manipulate the settings (much tidier, but a different screen, and a bit of a pain).
Now, the operator checks all the ColorRamp settings, then spawns additional value/RGB nodes and connects them to their respective sockets on the BL_ColorRamp in Unreal.
This also paves the way for supporting adding values directly on the Principled BSDF, but that has not been worked out quite yet.

Sooooooooo, I haven't done a ton of testing, but I think ColorRamps are now fully supported (minus the banding issues in the function, itself).