Values on Principled BSDF now auto-transfer
Thanks to the "Post Load Nodes", we now have any value (as in data type value, RGB is still coming) on the Principled BSDF node being copied and created as a node in Unreal.
RGB and RGBA will follow tomorrow, or shortly, at least.
It's starting to really need some cleanup, but I want to finish with the current features before neatening code.
I should also adjust the Bake Noise Nodes function to check if the noise node has a vector input before replacing it with the UV map, maybe it doesn't need to be a UV map node, specifically. But, a separate issue, to be dealt with after ColorRamp and Post Load Nodes are finished.