Principled BSDF data transfer cleanup
OK, so I was gathering Value, Vector and RGBA data from the Principled BSDF to recreate in Unreal. Started getting nodes with no values, not connected to anything.
Turns out, the Vector inputs on the Principled BSDF are only for Normal, Clearcoat Normal and Tangent data AKA data that you won't be typing in by hand.
So, I removed the Vector data from the pass, and tidied up the offsets for the "Post Load Nodes" that carry all the Principled BSDF or ColorRamp settings.
All necessary data should now transfer without the need for additional nodes in Blender!