Texture Coord bug fix, Bump node support, sort of
Texture Coordinate nodes threw an error, now, all of their outputs are being set to the single output of Unreal's TexCoord.
This node should probably receive the same attention as the Math, or Material Functions, and spawn different Unreal nodes, depending on which output is used, as Unreal uses separate nodes for Camera vs Object data etc. That will be another update.

Added support for Normal and NormalMap nodes, in the form of Reroute nodes, if there are better approximations of the functions of Normal and NormalMap in Unreal then I'll swap them in.

Added Bump node support, and worked out bugs, only to find that Unreal's NormalFromHeightmap node won't accept an RGB input, only T2D!
Hmmm, but it's a TEXTURE sample that's outputting the RGB, surely there's a way to get a TEXTURE2D output from a TEXTURE!?!??!
No, there is not.
The workaround for me, currently, is to create a Texture OBJECT, plug that into the Texture input of the Texture SAMPLE, then also plug the output of the Texture OBJECT into the NormalFromHeightmap node, but this then demands a Texture Coordinate to keep it aligned to the RGB output of the Texture Sample. Fuck.

In theory, I could add Texture Objects to the Post Load Nodes pass, I'd just have to check to make sure that approach isn't double loading texture data. But that does also pose some problems for making the connection, and I'm not sure how much that particular material function actually sees use in games.

This also lead to me finding a number of bugs:

1 - packed images retain their filepath from the original creator's PC - after downloading a number of files from Blendswap to test, I found that all the .blends with packed images were failing with error messages indicating the texture were stored in a location that didn't exist on my PC, leading me to the above conclusion. I could maybe unpack the images first, to avoid non-existent filepaths, but I'm not sure what's involved in that process, so it could take time to figure out, or if it's worth it.

2 - Mapping node is just a buggy pile of nonsense. Initial test were promising, but either I made some mistakes in modifying it, or the transition from 3D to 2D has a larger effect than I'd hoped. The node setup will recreate the same results in Blender as the Mapping node, but, in Unreal, it gives inconsistent/unpredictable results, and takes so damn long to compile on my machine that it also make me question whether it's worth pursuing further.